requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
python-dotenv>=1.0.0

# Object Detection & Tracking
ultralytics>=8.0.0
supervision>=0.16.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
mplsoccer>=1.2.0

# Dashboard
streamlit>=1.28.0
streamlit-drawable-canvas>=0.9.0

# AI / LLM
openai>=1.0.0
google-generativeai>=0.3.0

# API Client
requests>=2.31.0

# Server (optional - for local server mode)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: GPU acceleration
# torch>=2.0.0
# torchvision>=0.15.0
